+++ packages/kanban-lite/src/shared/config.ts
@@ -1,3 +1,4 @@
 export * from './config/types'
 export * from './config/io'
 export * from './config/normalize'
+export { withConfigReadCache, invalidateConfigReadCache } from './config/readCache'